I did a few things to get around https://github.com/ROCmSoftwarePlatform/Thrust/issues/8. The reason that the kernel launch was failing was because the runtime could not find the kernel:
_ZN6thrust6system4cuda6detail6detail23launch_closure_by_valueINS2_36commutative_reduce_intervals_closureINS_18transform_iteratorINS_6detail21predicate_to_integralI6is_oddIiElEENS7_15normal_iteratorINS_10device_ptrIiEEEElNS_11use_defaultEEENSC_INS_7pointerIlNS2_3tagESG_SG_EEEENS_4plusIlEENS0_6detail8internal21uniform_decompositionIlEENS3_20blocked_thread_arrayEEEEEvT_

This kernel was not being compiled as it was behind #ifdef guards (THRUST_DEVICE_COMPILER == THRUST_DEVICE_COMPILER_NVCC). However, I enabled THRUST_DEVICE_COMPILER_NVCC preprocessor macro if we were using HCC to compile HIP device code. The only negative side effect was that I needed to disable some code in ../thrust/system/cuda/detail/detail/alignment.h.

The other problem that I encountered was that the launch parameters to commutative_reduce_intervals_closure were giving a block_size = 0. For this patch, I hardcoded the block_size = 256, and max_blocks = 64. THIS IS A TEMPORARY HACK and should be FIXED!!! There is a bug in the launch calculator code in the default_block_configuration() method. I also added a hack to fix the histogram.cu example. Same issue with the adjacent_difference_closure kernel launch paramters.